# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DecisionNode {
  /** @minLength 1 */
  id: string;
  /** @minLength 1 */
  label: string;
  /** @minLength 1 */
  prompt: string;
  context?: string;
  terminalOutcome?: string;
}

export type DecisionEdgeBranch = typeof DecisionEdgeBranch[keyof typeof DecisionEdgeBranch];


export const DecisionEdgeBranch = {
  YES: 'YES',
  NO: 'NO',
} as const;

export interface DecisionEdge {
  id: string;
  source: string;
  target: string;
  branch: DecisionEdgeBranch;
}

export interface DecisionGraph {
  nodes: DecisionNode[];
  edges: DecisionEdge[];
}

export interface GraphValidation {
  valid: boolean;
  errors: string[];
  warnings: string[];
}

export type ExecutionInputMode = typeof ExecutionInputMode[keyof typeof ExecutionInputMode];


export const ExecutionInputMode = {
  stub: 'stub',
  ai: 'ai',
} as const;

export interface ExecutionInput {
  graph: DecisionGraph;
  startNodeId: string;
  input?: string;
  mode?: ExecutionInputMode;
}

export type ExecutionLogEntryStatus = typeof ExecutionLogEntryStatus[keyof typeof ExecutionLogEntryStatus];


export const ExecutionLogEntryStatus = {
  running: 'running',
  routed: 'routed',
  completed: 'completed',
  error: 'error',
} as const;

export type ExecutionLogEntryDecision = typeof ExecutionLogEntryDecision[keyof typeof ExecutionLogEntryDecision];


export const ExecutionLogEntryDecision = {
  YES: 'YES',
  NO: 'NO',
} as const;

export interface ExecutionLogEntry {
  sequence: number;
  nodeId: string;
  label: string;
  status: ExecutionLogEntryStatus;
  decision?: ExecutionLogEntryDecision;
  message: string;
  latencyMs?: number;
}

export type ExecutionResultStatus = typeof ExecutionResultStatus[keyof typeof ExecutionResultStatus];


export const ExecutionResultStatus = {
  completed: 'completed',
  failed: 'failed',
} as const;

export interface ExecutionResult {
  status: ExecutionResultStatus;
  executionId: string;
  terminalNodeId?: string;
  terminalOutcome?: string;
  visitedNodeIds: string[];
  logs: ExecutionLogEntry[];
  error?: string;
}

export interface ExecutionCapabilities {
  inngestRegistered: boolean;
  modes: string[];
  maxSteps: number;
}