# lib/api-zod/src/generated/types/decisionEdge.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DecisionEdgeBranch } from './decisionEdgeBranch';

export interface DecisionEdge {
  id: string;
  source: string;
  target: string;
  branch: DecisionEdgeBranch;
}