# lib/api-zod/src/generated/types/decisionGraph.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DecisionEdge } from './decisionEdge';
import type { DecisionNode } from './decisionNode';

export interface DecisionGraph {
  nodes: DecisionNode[];
  edges: DecisionEdge[];
}