# lib/api-zod/src/generated/types/decisionNode.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DecisionNode {
  /** @minLength 1 */
  id: string;
  /** @minLength 1 */
  label: string;
  /** @minLength 1 */
  prompt: string;
  context?: string;
  terminalOutcome?: string;
}