# lib/api-zod/src/generated/types/executionInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DecisionGraph } from './decisionGraph';
import type { ExecutionInputMode } from './executionInputMode';

export interface ExecutionInput {
  graph: DecisionGraph;
  startNodeId: string;
  input?: string;
  mode?: ExecutionInputMode;
}