# lib/api-zod/src/generated/types/executionLogEntry.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ExecutionLogEntryDecision } from './executionLogEntryDecision';
import type { ExecutionLogEntryStatus } from './executionLogEntryStatus';

export interface ExecutionLogEntry {
  sequence: number;
  nodeId: string;
  label: string;
  status: ExecutionLogEntryStatus;
  decision?: ExecutionLogEntryDecision;
  message: string;
  latencyMs?: number;
}