# lib/api-zod/src/generated/types/executionResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ExecutionLogEntry } from './executionLogEntry';
import type { ExecutionResultStatus } from './executionResultStatus';

export interface ExecutionResult {
  status: ExecutionResultStatus;
  executionId: string;
  terminalNodeId?: string;
  terminalOutcome?: string;
  visitedNodeIds: string[];
  logs: ExecutionLogEntry[];
  error?: string;
}