# lib/api-zod/src/generated/types/graphValidation.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GraphValidation {
  valid: boolean;
  errors: string[];
  warnings: string[];
}